feat(condo) INFRA-91: learn how to send traces manually
