+++ apps/condo/index.js
@@ -23,6 +23,7 @@ const FileAdapter = require('@condo/domains/common/utils/fileAdapter')
 const { VersioningMiddleware } = require('@condo/domains/common/utils/VersioningMiddleware')
 const { UserExternalIdentityMiddleware } = require('@condo/domains/user/integration/UserExternalIdentityMiddleware')
 const { OIDCMiddleware } = require('@condo/domains/user/oidc')
+const {TracingMiddleware} = require("@open-condo/keystone/tracing");
 
 
 dayjs.extend(duration)
@@ -127,45 +128,6 @@ const checks = [
     getRedisHealthCheck(),
     getPostgresHealthCheck(),
 ]
-
-class TracingMiddleware {
-    async prepareMiddleware ({ keystone }) {
-        const app = express()
-
-        const opentelemetry = require('@opentelemetry/sdk-node')
-        const {
-            getNodeAutoInstrumentations,
-        } = require('@opentelemetry/auto-instrumentations-node')
-        const {
-            OTLPTraceExporter,
-        } = require('@opentelemetry/exporter-trace-otlp-proto')
-        const {
-            OTLPMetricExporter,
-        } = require('@opentelemetry/exporter-metrics-otlp-proto')
-        const { PeriodicExportingMetricReader } = require('@opentelemetry/sdk-metrics')
-
-        const sdk = new opentelemetry.NodeSDK({
-            traceExporter: new OTLPTraceExporter({
-                // optional - default url is http://localhost:4318/v1/traces
-                //url: 'http://0.0.0.0:4317',
-                // optional - collection of custom headers to be sent with each request, empty by default
-                headers: {},
-            }),
-            metricReader: new PeriodicExportingMetricReader({
-                exporter: new OTLPMetricExporter({
-                    //url: '<your-otlp-endpoint>/v1/metrics', // url is optional and can be omitted - default is http://localhost:4318/v1/metrics
-                    //url: 'http://0.0.0.0:4317',
-                    headers: {}, // an optional object containing custom headers to be sent with each request
-                    concurrencyLimit: 1, // an optional limit on pending requests
-                }),
-            }),
-            instrumentations: [getNodeAutoInstrumentations()],
-        })
-        sdk.start()
-
-        return app
-    }
-}
 
 const lastApp = conf.NODE_ENV === 'test' ? undefined : new NextApp({ dir: '.' })
 const apps = () => {
