+++ apps/condo/package.json
@@ -70,6 +70,8 @@
     "@opentelemetry/auto-instrumentations-node": "^0.38.0",
     "@opentelemetry/exporter-metrics-otlp-proto": "^0.41.1",
     "@opentelemetry/exporter-trace-otlp-proto": "^0.41.1",
+    "@opentelemetry/instrumentation-express": "^0.33.0",
+    "@opentelemetry/instrumentation-http": "^0.41.2",
     "@opentelemetry/sdk-metrics": "^1.15.1",
     "@opentelemetry/sdk-node": "^0.41.1",
     "@tinymce/tinymce-react": "^4.1.0",
